[Android] Added a base bottom sheet dialog fragment. Workaround for the landscape height bug
